fix (mappers/github/Repository): Repo Class deleted, Repository Class takes over its role
