+++ src/main/java/dev/orion/revision/chain/github/MoodleUserChecker.java
@@ -56,7 +56,7 @@ public class MoodleUserChecker extends AbstractChecker implements Checker {
 
         if (mUsers != null &&  gUser != null){
             // Verifies if the Moodle and Github name are the same
-            if (mUsers.getFirstUserName().equalsIgnoreCase(gUser.getName())) {
+               if (mUsers.getFirstUserName().equalsIgnoreCase(gUser.getName())) {
                 LOGGER.info("O usuário do Moodle e Github são os mesmos");
                 // send to the next checker
                 result = this.getNextChecker().check(input);
